fix: Fix case classes serialization to match API
